+++ src/index.html
@@ -147,5 +147,6 @@
         </p>
       </div>
     </footer>
+    <div class="popup-container"></div>
   </body>
 </html>
