+++ src/js/logOutFetch.js
@@ -1,0 +1,11 @@
+
+import fetchFunctions from './fetchMe';
+import renderCategories from '../templates/category.hbs';
+import renderFilter from '../templates/filter.hbs';
+import renderList from '../templates/watchAll.hbs';
+
+
+import onPaginationPage from './paginationCategories'
+
+
+
