+++ src/js/modalClass.js
@@ -6,16 +6,24 @@ import myAdvert from '../templates/my-advert.hbs';
 import selectedAdvert from '../templates/selected-advert.hbs';
 import authorizationModal from '../templates/authorization-modal.hbs';
 import popupSearch from '../templates/pop-up-search.hbs';
+import { openModalFind } from '../js/openSearchModal';
+
+
+
 
 const hbsFunctions = [
   renderCardlist,
   renderAddCard,
+  openModalFind,
   popupSearch,
   popupExitConfirm,
   myOffice,
   myAdvert,
   selectedAdvert,
   authorizationModal,
+  
+
+
 ];
